+++ test/controllers/hooks.test.js
@@ -12,7 +12,7 @@
 
 /* eslint-env mocha */
 
-import chai from 'chai';
+import { use, expect } from 'chai';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
 import nock from 'nock';
@@ -20,8 +20,7 @@ import { Blocks, Elements, Message } from 'slack-block-builder';
 import HooksController from '../../src/controllers/hooks.js';
 import { SiteDto } from '../../src/dto/site.js';
 
-chai.use(sinonChai);
-const { expect } = chai;
+use(sinonChai);
 
 function getExpectedSlackMessage(baseURL, channel, source, hlxConfig) {
   const cdnConfigPart = hlxConfig
@@ -117,13 +116,33 @@ describe('Hooks Controller', () => {
       expect(resp.status).to.equal(404);
       expect(slackClient.postMessage.notCalled).to.be.true;
     });
+
+    it('return 400 if hlx version is not an integer', async () => {
+      context.params = { hookSecret: 'hook-secret-for-cdn' };
+      context.data = { hlxVersion: 'not-integer' };
+
+      const resp = await hooksController.processCDNHook(context);
+      expect(resp.status).to.equal(400);
+      expect(slackClient.postMessage.notCalled).to.be.true;
+    });
+
+    it('return 400 if requestXForwardedHost has no text', async () => {
+      context.params = { hookSecret: 'hook-secret-for-cdn' };
+      context.data = { hlxVersion: 4, requestXForwardedHost: '' };
+
+      const resp = await hooksController.processCDNHook(context);
+      expect(resp.status).to.equal(400);
+      expect(slackClient.postMessage.notCalled).to.be.true;
+    });
   });
 
   describe('URL sanitization checks', () => {
     async function assertInvalidCase(input) {
       context.params = { hookSecret: 'hook-secret-for-cdn' };
       context.data = {
-        forwardedHost: input,
+        hlxVersion: 4,
+        requestPath: '/test',
+        requestXForwardedHost: input,
       };
       const resp = await (await hooksController.processCDNHook(context)).json();
       expect(resp).to.equal('CDN site candidate disregarded');
@@ -220,7 +239,8 @@ describe('Hooks Controller', () => {
         .replyWithError({ code: 'ECONNREFUSED', syscall: 'connect', message: 'rainy weather' });
 
       context.data = {
-        forwardedHost: 'some-domain.com, some-fw-domain.com',
+        hlxVersion: 4,
+        requestXForwardedHost: 'some-domain.com, some-fw-domain.com',
       };
 
       const resp = await (await hooksController.processCDNHook(context)).json();
@@ -235,7 +255,8 @@ describe('Hooks Controller', () => {
         .reply(200, invalidHelixDom);
 
       context.data = {
-        forwardedHost: 'some-domain.com, some-fw-domain.com',
+        hlxVersion: 4,
+        requestXForwardedHost: 'some-domain.com, some-fw-domain.com',
       };
 
       const resp = await (await hooksController.processCDNHook(context)).json();
@@ -252,7 +273,9 @@ describe('Hooks Controller', () => {
         .reply(200, validHelixDom);
 
       context.data = {
-        forwardedHost: 'some-domain.com, some-fw-domain.com',
+        hlxVersion: 4,
+        requestPath: '/test',
+        requestXForwardedHost: 'some-domain.com, some-fw-domain.com',
       };
       context.params = { hookSecret: 'hook-secret-for-cdn' };
     });
@@ -309,7 +332,8 @@ describe('Hooks Controller', () => {
       context.dataAccess.siteCandidateExists.resolves(false);
       context.dataAccess.upsertSiteCandidate.resolves();
       context.data = {
-        forwardedHost: 'some-domain.com, main--some-site--some-owner.hlx.live',
+        hlxVersion: 5,
+        requestXForwardedHost: 'some-domain.com, main--some-site--some-owner.hlx.live',
       };
 
       const hlxConfig = {
@@ -390,7 +414,8 @@ describe('Hooks Controller', () => {
     it('CDN candidate is processed and slack message sent', async () => {
       const hlx5Config = { cdn: { prod: { host: 'some-cdn-host.com' } } };
       context.data = {
-        forwardedHost: 'some-domain.com, some-fw-domain.com, main--some-site--some-owner.hlx.live',
+        hlxVersion: 5,
+        requestXForwardedHost: 'some-domain.com, some-fw-domain.com, main--some-site--some-owner.hlx.live',
       };
       context.params = { hookSecret: 'hook-secret-for-cdn' };
 
@@ -425,9 +450,27 @@ describe('Hooks Controller', () => {
       expect(actualMessage).to.deep.equal(expectedMessage);
     });
 
+    it('CDN candidate is processed when fetching config is skipped for hlx version < 5', async () => {
+      context.data = {
+        hlxVersion: 4,
+        requestXForwardedHost: 'some-domain.com, some-fw-domain.com, main--some-site--some-owner.hlx.live',
+      };
+      context.params = { hookSecret: 'hook-secret-for-cdn' };
+
+      nock('https://some-cdn-host.com')
+        .get('/')
+        .reply(200, validHelixDom);
+
+      const resp = await (await hooksController.processCDNHook(context)).json();
+
+      expect(context.log.info).to.have.been.calledWith('HLX version is 4. Skipping fetching hlx config');
+      expect(resp).to.equal('CDN site candidate is successfully processed');
+    });
+
     it('CDN candidate is processed even with hlx config 404', async () => {
       context.data = {
-        forwardedHost: 'some-domain.com, some-fw-domain.com, main--some-site--some-owner.hlx.live',
+        hlxVersion: 5,
+        requestXForwardedHost: 'some-domain.com, some-fw-domain.com, main--some-site--some-owner.hlx.live',
       };
       context.params = { hookSecret: 'hook-secret-for-cdn' };
 
@@ -447,7 +490,8 @@ describe('Hooks Controller', () => {
 
     it('CDN candidate is processed even with error status for helix config request', async () => {
       context.data = {
-        forwardedHost: 'some-domain.com, some-fw-domain.com, main--some-site--some-owner.hlx.live',
+        hlxVersion: 5,
+        requestXForwardedHost: 'some-domain.com, some-fw-domain.com, main--some-site--some-owner.hlx.live',
       };
       context.params = { hookSecret: 'hook-secret-for-cdn' };
 
@@ -467,7 +511,8 @@ describe('Hooks Controller', () => {
 
     it('CDN candidate is processed even when fetch throws for helix config request', async () => {
       context.data = {
-        forwardedHost: 'some-domain.com, some-fw-domain.com, main--some-site--some-owner.hlx.live',
+        hlxVersion: 5,
+        requestXForwardedHost: 'some-domain.com, some-fw-domain.com, main--some-site--some-owner.hlx.live',
       };
       context.params = { hookSecret: 'hook-secret-for-cdn' };
 
@@ -487,7 +532,8 @@ describe('Hooks Controller', () => {
 
     it('CDN candidate is processed and slack message sent even if site was added previously but is not aem_edge', async () => {
       context.data = {
-        forwardedHost: 'some-domain.com, some-fw-domain.com',
+        hlxVersion: 4,
+        requestXForwardedHost: 'some-domain.com, some-fw-domain.com',
       };
       context.params = { hookSecret: 'hook-secret-for-cdn' };
       context.dataAccess.getSiteByBaseURL.resolves(SiteDto.fromJson({
@@ -518,7 +564,8 @@ describe('Hooks Controller', () => {
 
     it('Slack message sending fails for CDN candidate', async () => {
       context.data = {
-        forwardedHost: 'some-domain.com, some-fw-domain.com',
+        hlxVersion: 5,
+        requestXForwardedHost: 'some-domain.com, some-fw-domain.com',
       };
       context.params = { hookSecret: 'hook-secret-for-cdn' };