+++ test/support/s3.test.js
@@ -12,14 +12,13 @@
 
 /* eslint-env mocha */
 
-import chai from 'chai';
+import { use, expect } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import sinon from 'sinon';
 
 import { s3ClientWrapper } from '../../src/support/s3.js';
 
-chai.use(chaiAsPromised);
-const { expect } = chai;
+use(chaiAsPromised);
 
 describe('S3 client wrapper tests', () => {
   let mockRequest;
@@ -32,6 +31,9 @@ describe('S3 client wrapper tests', () => {
       log: console,
       runtime: {
         region: 'us-east-1',
+      },
+      env: {
+        S3_BUCKET_NAME: 'test-bucket',
       },
     };
 
@@ -57,6 +59,7 @@ describe('S3 client wrapper tests', () => {
     expect(firstCall.args[1].s3).to.be.an('object');
     expect(firstCall.args[1].s3.s3Client).to.be.an('object');
     expect(firstCall.args[1].s3.getSignedUrl).to.be.a('function');
+    expect(firstCall.args[1].s3.s3Bucket).to.equal(mockContext.env.S3_BUCKET_NAME);
   });
 
   it('does not create a new S3Client if one already exists in the context', async () => {